Refactor ScanComponentsButton
